requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
# Editable install with no version control (datapizza-ai-clients-anthropic==0.0.3)
-e /Users/nicolagnasso/projects/DnD/datapizza-ai-main/datapizza-ai-clients/datapizza-ai-clients-anthropic
# Editable install with no version control (datapizza-ai-clients-google==0.0.4)
-e /Users/nicolagnasso/projects/DnD/datapizza-ai-main/datapizza-ai-clients/datapizza-ai-clients-google
# Editable install with no version control (datapizza-ai-clients-openai==0.0.6)
-e /Users/nicolagnasso/projects/DnD/datapizza-ai-main/datapizza-ai-clients/datapizza-ai-clients-openai
# Editable install with no version control (datapizza-ai-clients-openai-like==0.0.6)
-e /Users/nicolagnasso/projects/DnD/datapizza-ai-main/datapizza-ai-clients/datapizza-ai-clients-openai-like
# Editable install with no version control (datapizza-ai-core==0.0.6)
-e /Users/nicolagnasso/projects/DnD/datapizza-ai-main/datapizza-ai-core
# Editable install with no version control (datapizza-ai-embedders-openai==0.0.5)
-e /Users/nicolagnasso/projects/DnD/datapizza-ai-main/datapizza-ai-embedders/openai
# Editable install with no version control (datapizza-ai-vectorstores-qdrant==0.0.4)
-e /Users/nicolagnasso/projects/DnD/datapizza-ai-main/datapizza-ai-vectorstores/datapizza-ai-vectorstores-qdrant
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.121.0
google-auth==2.43.0
google-genai==1.49.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.1
jsonref==1.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.3.4
openai==2.7.1
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
portalocker==3.2.0
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.0.1
PyYAML==6.0.3
qdrant-client==1.15.1
requests==2.32.5
rich==14.2.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.49.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
zipp==3.23.0
